+++ ui/lib/audio_engine_web.dart
@@ -348,6 +348,9 @@ class AudioEngine implements AudioEngineInterface {
   String setAudioClipWarp(int trackId, int clipId, bool warpEnabled, double stretchFactor, int warpMode) => 'OK';
 
   @override
+  String setAudioClipTranspose(int trackId, int clipId, int semitones, int cents) => 'OK';
+
+  @override
   void removeAudioClip(int trackId, int clipId) {}
 
   @override